factor out convert module
