+++ pyproject.toml
@@ -1,6 +1,5 @@
 [build-system]
 requires = [
-    "setuptools >= 35.0.2",
-    "setuptools_scm >= 2.0.0, <3"
+    "setuptools >= 46.4.0"
 ]
 build-backend = "setuptools.build_meta"
