+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = runcrate
-version = 0.1.0
+version = attr: runcrate.__version__
 license = Apache License, version 2.0 (Apache-2.0)
 description = Workflow Run RO-Crate toolkit
 long_description = Runcrate is a software toolkit to manipulate `Workflow Run RO-Crate <https://www.researchobject.org/workflow-run-crate/>`_ packages.
