+++ tests/test_step_mapping.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from runcrate import ProvCrateBuilder, get_workflow
+from runcrate.convert import ProvCrateBuilder, get_workflow
 
 
 def test_step_maps(data_dir):
